requirements.txt
streamlit>=1.24.0
psycopg2-binary>=2.9.6
pandas>=1.5.3
plotly>=5.13.1
python-dotenv>=1.0.0
requests>=2.31.0
sqlalchemy>=2.0.0